Merge commit '9f29e1223cd5000a7429ca63ef47a7033b4cdbf7' into 9-develop

# Conflicts:
#	app/build.gradle

+++ app/src/main/java/com/think/runex/java/Pages/RecordPage/v2/RecordPage.java
@@ -1301,7 +1301,7 @@ public class RecordPage extends xFragment implements OnMapReadyCallback
         final String mtn = ct + "apiWorkOuts() ";
         final rqAddWorkOutsHistory request = new rqAddWorkOutsHistory();
 
-        final double recordTime = recTimeAsSec(currentRecorder.durationMillis);
+        final long recordTime = recTimeAsSec(currentRecorder.durationMillis);
         final double recordPace = paceAsMin(currentRecorder.paceMillis);
 
         //request.setActivity_type(Globals.ACTIVITY_RUN);
